+++ src/config/deserialized_account_config.rs
@@ -45,7 +45,7 @@ macro_rules! make_account_config {
             pub signature: Option<String>,
             /// Overrides the signature delimiter for this account.
             pub signature_delimiter: Option<String>,
-            /// Overrides the default page size for this account.
+	    /// Overrides the default page size for this account.
             pub default_page_size: Option<usize>,
             /// Overrides the notify command for this account.
             pub notify_cmd: Option<String>,
@@ -56,6 +56,10 @@ macro_rules! make_account_config {
 	    /// Represents the text/plain format as defined in the
 	    /// [RFC2646](https://www.ietf.org/rfc/rfc2646.txt)
             pub format: Option<Format>,
+            /// Overrides the default headers displayed at the top of
+            /// the read message.
+	    #[serde(default)]
+            pub read_headers: Vec<String>,
 
             /// Makes this account the default one.
             pub default: Option<bool>,
@@ -102,6 +106,7 @@ macro_rules! make_account_config {
             	    notify_query: self.notify_query.clone(),
             	    watch_cmds: self.watch_cmds.clone(),
             	    format: self.format.clone(),
+		    read_headers: self.read_headers.clone(),
 
             	    default: self.default.clone(),
             	    email: self.email.clone(),